requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.0.3
django_coverage_plugin==1.2.2
Sphinx
django-extensions==1.6.1
Werkzeug==0.11.3
django-test-plus==1.0.11
factory_boy==2.6.0

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.4

# improved REPL
ipdb==0.8.1

# pytest!
pytest==2.9.0
pytest-django==2.9.1
pytest-sugar==0.5.1
pytest-cov==2.2.1
pytest-factoryboy==1.1.5
